docs: pastas reorganizadas (UserService agora é UserRepository)
