+++ userApi/src/main/java/com/github/userApi/dto/IncluirUserRequest.java
@@ -1,0 +1,12 @@
+package com.github.userApi.dto;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.Data;
+
+@Data
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class IncluirUserRequest {
+    private Long id;
+    private String name;
+    private int age;
+}
