+++ userApi/src/main/java/com/github/userApi/repository/UserRepository.java
@@ -1,0 +1,23 @@
+package com.github.userApi.repository;
+import java.util.List;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserRepository {
+//    private final UserRepository userRepository;
+//
+//    public ClienteService(UserRepository clienteRepository){
+//        this.UserRepository = clienteRepository;
+//    }
+//
+//    public List<User> listar(){
+//        return UserRepository.FindAll();
+//    }
+//
+//    public User getUser(Long id){
+//        return userRepository.findById(id)
+//                //.orElseThrow(() -> new UserNotFoundException("User não localizado" + id));
+//    }
+//
+}
